[add/update] folder structure and refactor app
