+++ .gitignore
@@ -1,0 +1,18 @@
+# Python files
+__pycache__/
+*.pyc
+*.pyo
+
+# Logs
+*.log
+
+# IDE-specific files
+.vscode/
+.idea/
+
+# Environment-specific files
+.env
+venv/
+test/replays/
+config.py
+SC2_sounds.json
